Backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.26.0

# LLM & AI
langchain==0.1.0
langchain-openai==0.0.5

# Voice
elevenlabs==0.2.24

# Optional: For NVIDIA Endpoints
# langchain-nvidia-ai-endpoints==0.0.1
